Add .gitnutzrc.json and Update Go Source Files
Several updates were made to the Go codebase and a new configuration
file named `.gitnutzrc.json` was added with contents specifying the
name "Jarbles Framework".

In `common.go`, imports were added for `encoding/json`, `fmt`, and
`time` packages. New functions `PayloadParse`, `PayloadMustParse`,
`PayloadGetString`, and `SleepAtLeast` were introduced. The `slugify`
function was updated to use `strings.ReplaceAll` instead of
`strings.Replace`, along with minor code adjustments like added
comments and indentation changes.

The `extension.go` file saw the introduction of the "context" package
import, a new `NoLayout` boolean field in the `ExtensionResponse`
struct, and modifications in logging statements to a more context-aware
form. The `Execute` function was updated to accommodate these new
logging methods.

Changes in `logger.go` include the replacement of "strconv" and
"strings" imports with "time" and adjustments to logging functionality.
The "trunc" variable was changed to "pretty" and its type was adjusted
from int to bool, affecting how attributes are handled if pretty logging
is enabled. The `WithAttrs` function has been marked as unimplemented,
indicating a work-in-progress or planned feature disablement.